TT-1299: Audit tool to see users in all deployers
This is a first pass at a tool to audit and present the state of accounts on
our jenkins instances, currently limited to deployers.

There's many ideas for how to improve on this, including:

- add ci and smoketester boxes to the checked list
- tie this into the users.json

Next steps would include:

- add a script to delete users from jenkins
- add a script to grant users admin privileges (without editing config.xml)

It's in this repo because we need a recent version of ruby (compared to
verify-puppet), and this repo seems a good place to converse user account
related things.

Pair: @javindo @jinnko