bau: rescue a cryptic error
If the terminal isn't wide enough the tty-table gem emits a NoMethodFound error
even though that seems to have nothing to do with what's happening.  We handle
the exception and give the user a hint to make their terminal wider.  This is
the quick workaround.  A fix might be to exclude empty columns, but that's work.

Thanks to @tlwr for figuring this out.

Pairless