BAU: add gds names and preferred name
gds name is the user's email prefix - taken from their gds email,
and useful in future for Slack naming.

preferred name is just the first name in their "Names" list for now,
added so scripts can choose a name to use to address people, without
assuming that the first listed Name is somehow important

Readme is also updated with the current meanings of these fields,
and plans to deprecate the slack id and slack name fields.