BAU: Define repo and team Terraform modules
The `verify_repo` module defines a Github repository with the necessary
permissions and branch protection settings. `master_users` defines the
users who have permission to merge PRs and push directly to the master
branch. `push_teams` defines the teams that have the ability to push
branches and raise PRs against the repo.

The `verify_team` module defines a team within the organisation. At the
moment we have to use simple team names (lowercase and hyphenated) so
that they match their slugs.

Author: @vixus0