BAU: Allow RE team to have access to Core Hub
At this moment, RE team are not able to create a branch and create a
pull request on Core Hub. This commit allows them to have access to
scripts for building debian packages and upstart files for starting and
stopping SAML SOAP Proxy, Policy, etc. This is needed for working on
Ubuntu upgrade to 18.04 LTS.

Author: @adityapahuja