Make `ida-build-bot` have admin access on `verify-event-infrastructure` repo.
This is needed in order for our Concourse pipeline, which runs Terraform plan on Pull Requests (PR), to be able to
update PR status.

The Concourse pipeline uses the `github-pr` resource (https://github.com/telia-oss/github-pr-resource).
Since this repo is private, the Concourse resource needs admin access, as mentioned in its docs.